# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "todo",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/api/ai/prompt": {
      "get": {
        "summary": "askToAI",
        "deprecated": false,
        "description": "AI에게 질문을 보내고 응답을 받습니다.",
        "tags": [
          "ChatAIController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "xetpof"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-ai-prompt-GET",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "AI에게 보낼 질문",
            "required": true,
            "example": "hello",
            "name": "question"
          }
        ]
      }
    },
    "/api/ai/task": {
      "get": {
        "summary": "createTaskToAI",
        "deprecated": false,
        "description": "제공된 프롬프트를 사용하여 AI로 생성된 Task 를 제시합니다.",
        "tags": [
          "ChatAIController"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTaskFromAIResponse"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"title\": \"pass test\",\n  \"description\": \"programming todo app\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-ai-task-GET",
        "parameters": [
          {
            "schema": {
              "format": "string",
              "type": "string"
            },
            "in": "query",
            "description": "AI에게 Task 를 생성해달라고 요청",
            "required": false,
            "example": "Add Task. Title: pass test, Description: programming todo app",
            "name": "prompt"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTaskFromAIResponse": {
        "properties": {
          "title": {
            "format": "string",
            "description": "Task 제목",
            "type": "string",
            "example": "Pass Test"
          },
          "description": {
            "format": "string",
            "description": "Task 설명",
            "type": "string",
            "example": "Programming Todo App"
          }
        }
      },
...
    }
  }
}